moved addTodo to it's own module
