+++ index.html
@@ -12,7 +12,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>TodoMVC</title>
     <link rel="stylesheet" href="./css/style.css">
-    <script src="./scripts/main.js" defer></script>
+    <script type="module" src="./scripts/main.js" defer></script>
 </head>
 
 <body>
