+++ scripts/main.js
@@ -1,5 +1,6 @@
+import { addTodo } from "./todo.js";
+
 const form = document.getElementById("form");
-const main = document.getElementById("main");
 
 form.addEventListener("submit", (e) => {
   e.preventDefault();
@@ -8,24 +9,3 @@ form.addEventListener("submit", (e) => {
   addTodo(term.value);
   term.value = "";
 });
-
-function addTodo(todo) {
-  const todoContainer = document.createElement("div");
-
-  const todoInput = document.createElement("input");
-  todoInput.type = "checkbox";
-  todoInput.checked = false;
-
-  const todoLabel = document.createElement("label");
-  todoLabel.textContent = todo;
-
-  todoInput.onchange = () => {
-    if (todoInput.checked) {
-        todoInput.classList.add("checked");
-    } else {
-        todoInput.classList.remove("checked")
-    }
-  }
-  todoContainer.append(todoInput, todoLabel);
-  main.append(todoContainer);
-}
